GameView changes
Can now navigate to the hospital, power up den, villains lair and shop. With the abillity to move back to the map panel(homebase).

Changed the way the makemove() behaves, it now makes the player's move then will check if they have won if not makes the computers move then checks for a winner again.